+++ dep.go
@@ -77,8 +77,15 @@ func trimGoVersion(version string) (string, error) {
 	if len(p) < 2 {
 		return "", fmt.Errorf("Error determing major go version from: %q", version)
 	}
-	if strings.Contains(p[1], "beta") {
-		p[1] = strings.Split(p[1], "beta")[0]
+	var split string
+	switch {
+	case strings.Contains(p[1], "beta"):
+		split = "beta"
+	case strings.Contains(p[1], "rc"):
+		split = "rc"
+	}
+	if split != "" {
+		p[1] = strings.Split(p[1], split)[0]
 	}
 	return p[0] + "." + p[1], nil
 }